Исправлен баг с изменением SSH ключа
