+++ devicemanager/tests/test_connection_status.py
@@ -5,8 +5,8 @@ from unittest.mock import Mock, patch
 
 from django.test import SimpleTestCase
 
-from devicemanager.device_connector.connection_status import (
-    ConnectionStatusStore,
+from devicemanager.device_connector.connection_status import ConnectionStatusStore
+from devicemanager.device_connector.ssh_host_keys import (
     PendingSSHHostKeyChange,
     SSHKeyInfo,
     SSHKnownHostsStore,
@@ -151,7 +151,7 @@ class SSHKnownHostsStoreTests(SimpleTestCase):
             new_key_lines=("[192.0.2.10]:2222 ssh-ed25519 AAAAnew",),
         )
 
-    @patch("devicemanager.device_connector.connection_status.subprocess.run")
+    @patch("devicemanager.device_connector.ssh_host_keys.subprocess.run")
     def test_confirmation_appends_only_previously_scanned_key_lines(self, run):
         """Confirmation writes the exact pending key material without a rescan."""
 
@@ -171,7 +171,7 @@ class SSHKnownHostsStoreTests(SimpleTestCase):
             self.assertEqual(command[3], "-f")
             self.assertEqual(Path(command[4]).parent, known_hosts_path.parent)
 
-    @patch("devicemanager.device_connector.connection_status.subprocess.run")
+    @patch("devicemanager.device_connector.ssh_host_keys.subprocess.run")
     def test_accept_current_scans_and_writes_key_in_one_operation(self, run):
         """Automatic acceptance writes the key scanned inside the locked operation."""
 
@@ -215,7 +215,53 @@ class SSHKnownHostsStoreTests(SimpleTestCase):
                 "192.0.2.10 ssh-ed25519 AAAAnew\n",
             )
 
-    @patch("devicemanager.device_connector.connection_status.subprocess.run")
+    @patch("devicemanager.device_connector.ssh_host_keys.subprocess.run")
+    def test_accept_changed_uses_warning_when_keyscan_fails(self, run):
+        """Automatic acceptance removes the old key when only warning data is available."""
+
+        command_results = iter(
+            [
+                Mock(
+                    returncode=0,
+                    stdout="192.0.2.10 ssh-rsa AAAAold\n",
+                    stderr="",
+                ),
+                Mock(returncode=1, stdout="", stderr="keyscan failed"),
+                Mock(
+                    returncode=0,
+                    stdout="192.0.2.10 ssh-rsa AAAAold\n",
+                    stderr="",
+                ),
+                Mock(returncode=0, stdout="2048 SHA256:old host (RSA)\n", stderr=""),
+            ]
+        )
+
+        def run_command(command, **kwargs):
+            """Emulate warning inspection and old key removal."""
+
+            if command[:2] == ["ssh-keygen", "-R"]:
+                Path(command[4]).write_text("", encoding="utf-8")
+                return Mock(returncode=0, stdout="", stderr="")
+            return next(command_results)
+
+        run.side_effect = run_command
+        with TemporaryDirectory() as temporary_directory:
+            known_hosts_path = Path(temporary_directory) / ".ssh" / "known_hosts"
+            known_hosts_path.parent.mkdir()
+            known_hosts_path.write_text("192.0.2.10 ssh-rsa AAAAold\n", encoding="utf-8")
+            store = SSHKnownHostsStore(known_hosts_path)
+
+            change = store.accept_changed(
+                "192.0.2.10",
+                22,
+                self.make_change().detected_at,
+                ("The fingerprint for the ED25519 key sent by the remote host is\n" "SHA256:new."),
+            )
+
+            self.assertEqual(change.new_keys[0].fingerprint, "SHA256:new")
+            self.assertEqual(known_hosts_path.read_text(encoding="utf-8"), "")
+
+    @patch("devicemanager.device_connector.ssh_host_keys.subprocess.run")
     def test_inspect_warning_extracts_new_key_fingerprint(self, run):
         """OpenSSH warning provides a fallback fingerprint for manual confirmation."""
 
@@ -249,7 +295,7 @@ class SSHKnownHostsStoreTests(SimpleTestCase):
             self.assertEqual(change.new_keys[0].fingerprint, "SHA256:new")
             self.assertEqual(change.new_key_lines, ())
 
-    @patch("devicemanager.device_connector.connection_status.subprocess.run")
+    @patch("devicemanager.device_connector.ssh_host_keys.subprocess.run")
     def test_failed_confirmation_keeps_original_known_hosts(self, run):
         """A failed replacement never removes the currently trusted key."""
 
@@ -268,7 +314,7 @@ class SSHKnownHostsStoreTests(SimpleTestCase):
                 "192.0.2.10 ssh-ed25519 AAAAold\n",
             )
 
-    @patch("devicemanager.device_connector.connection_status.subprocess.run")
+    @patch("devicemanager.device_connector.ssh_host_keys.subprocess.run")
     def test_confirmation_without_scanned_lines_only_removes_old_key(self, run):
         """Warning-based confirmation lets the next SSH handshake save the new key."""
